Issue: WSW-17509 MISRA Rule 10.3 patch 11
[Issue]
Implicit conversion of datatypes.

[Fix]
Typecast with appropriate type

[Test]
Build success on IAR ARMGCC Keil MCUX.
Scan and Connect as STA.
uAP start and external station connect and ping.

[Open Issues/Todo]
Other 10.3 Issues.

[Reviewers]
Sandip Dalvi, Supriya Inamdar, Abhinav Kulkarni, Sanjay Yadav

+++ wifidriver/wifi-uap.c
@@ -249,7 +249,7 @@ int wifi_uap_prepare_and_send_cmd(mlan_private *pmpriv,
     (void)wifi_get_command_lock();
     HostCmd_DS_COMMAND *cmd = wifi_get_command_buffer();
 
-    cmd->seq_num = HostCmd_SET_SEQ_NO_BSS_INFO(0U /* seq_num */, 0U /* bss_num */, bss_type);
+    cmd->seq_num = HostCmd_SET_SEQ_NO_BSS_INFO(0U /* seq_num */, 0U /* bss_num */, (uint16_t)bss_type);
     cmd->result  = 0x0;
 
     mlan_status rv = wlan_ops_uap_prepare_cmd(pmpriv, cmd_no, cmd_action, cmd_oid, pioctl_buf, pdata_buf, cmd);
